refactor : PrizeEnum의 구조 변경 (description 참고)
매칭되는 번호 개수, 상금, 보너스 볼의 존재 여부
위의 3가지로 생성자를 변경
